* Exception Handling

* Cleanup unused files

* Quick fixes and better error handling

* Add OpenAPI specification

* Merge updates (#8)

* one csproj for all unittests

* shit aint working

* too stupid to do it

* suizid

* Added integration tests

---------


* Updated github workflow to work with new test setup

* Whoopsie

* Another whoopsie

* Whoopsie 3

---------

+++ DAL/Repositories/DocumentRepository.cs
@@ -1,23 +1,28 @@
 ﻿using DAL.Data;
 using DAL.Entities;
+using Logging;
 using Microsoft.EntityFrameworkCore;
 
 namespace DAL.Repositories;
 
-public class DocumentRepository(DocumentContext context) : IDocumentRepository
+public class DocumentRepository(DocumentContext context, ILogger logger) : IDocumentRepository
 {
     public async Task<IEnumerable<Document>> GetAllAsync()
     {
+        if (!CheckIfDocumentsSetIsInitialized()) return new List<Document>();
         return await context.Documents!.ToListAsync();
     }
 
-    public async Task<Document> GetByIdAsync(int id)
+    public async Task<Document?> GetByIdAsync(int id)
     {
-        return (await context.Documents!.FindAsync(id))!;
+        if (!CheckIfDocumentsSetIsInitialized()) return null;
+        return await context.Documents!.FindAsync(id);
     }
 
     public async Task<int> AddAsync(Document item)
     {
+        if (!CheckIfDocumentsSetIsInitialized()) return -1;
+        
         var entry = await context.Documents!.AddAsync(item);
         await context.SaveChangesAsync();
             
@@ -26,17 +31,30 @@ public class DocumentRepository(DocumentContext context) : IDocumentRepository
 
     public async Task UpdateAsync(Document item)
     {
+        if (!CheckIfDocumentsSetIsInitialized()) return;
+        
         context.Documents!.Update(item);
         await context.SaveChangesAsync();
     }
 
     public async Task DeleteAsync(int id)
     {
+        if (!CheckIfDocumentsSetIsInitialized()) return;
+        
         var item = await context.Documents!.FindAsync(id);
         if (item != null)
         {
             context.Documents.Remove(item);
             await context.SaveChangesAsync();
         }
+    }
+    
+    private bool CheckIfDocumentsSetIsInitialized()
+    {
+        if (context.Documents != null) return true;
+        
+        logger.Error("Documents set is not initialized.");
+        return false;
+
     }
 }
